third_party/blink/web_tests/external/wpt/dom/parts/basic-dom-part-declarative-brace-syntax-innerhtml.tentative-expected.txt

This is a testharness.js-based test.
[PASS] Declarative DOM Parts innerHTML <div><h1>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <div><h1 parseparts>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <div parseparts=""><h1>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <div parseparts=""><h1 parseparts>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <template><h1>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <template><h1 parseparts>content... (expect no parts)
[PASS] Declarative DOM Parts innerHTML <template parseparts="" class="expect_success"><h1>content... (expect parts)
[PASS] Declarative DOM Parts innerHTML <template parseparts="" class="expect_success"><h1 parseparts>content... (expect parts)
[FAIL] Basic NodePart parsing
  assert_equals: There should be one NodePart expected 1 but got 0
[PASS] Basic AttributePart cloning with values
Harness: the test ran to completion.
